requirements.txt
aioredis==2.0.1
channels==4.0.0
channels-redis==4.1.0
django==4.2
djangorestframework==3.14.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.14.0
django-filter==23.1
mysqlclient==2.1.1